Fix potential panic
The change [1] introduced a potential panic as it is trying to log the
app name when a failure happens, but the app variable is overridden
during such a failure and is thus nil.

This results in the following panic.

    panic: runtime error: invalid memory address or nil pointer dereference
    [signal SIGSEGV: segmentation violation code=0x1 addr=0x20 pc=0x20ddccb]

    goroutine 367 [running]:
    internal/pkg/argocd.generateDiffOfAComponent()
	    internal/pkg/argocd/argocd.go:477
    internal/pkg/argocd.GenerateDiffOfChangedComponents()
	    internal/pkg/argocd/argocd.go:561
    internal/pkg/githubapi.HandlePREvent()
	    internal/pkg/githubapi/github.go:161
    internal/pkg/githubapi.handleEvent()
	    internal/pkg/githubapi/github.go:382
    created by internal/pkg/githubapi.ReciveWebhook
	    internal/pkg/githubapi/github.go:322

By instead using the name from the query constructed prior, the panic
should be avoided.

[1] adbd91328b2edce48ee4607f101ecb2b6528d77b