Increase event handling context timeout
Recently some promotion pull requests have logged that they are reaching
the timeout of handling the request, causing updates of the Github
checks to get stuck in a pending state.

This increases the timeout to try to mitigate this issue.

One alternative considered was to use a separate context without a
timeout for setting the commit status, or having a separate timeout for
it. At least for the updating of the success/failure update. Given there
is extensive room for even more improvements through such refactorings,
increasing the timeouts is a quick mitigation that leaves additional
work for the future.