Refactor to recover from panic in event handlers
This change refactors the event handling logic such that a deferred
panic handler can log panics in the downstream handler logic. This
should avoid crashing when such panics occur, and instead it would log
the panic using the logger.

Additionally, the parsing of the event payload to determine which
handling logic to invoke is separated out, and now also indicates
whether a match was found. This is to allow PR status updates to be
applied once, when appropriate, and to enable ensuring that the
success/failure update is always applied.

To achieve this the individual downstream logic is factored out into
separate functions, and errors encountered in them are returned where
prHandleError were previously set.

Getting the default branch and Telefonistka config is duplicated in each
handler as needed.