Increase processing timeout
Calculating diffs on promotion pull requests like [1] fails since the
context is timing out and processing is aborted. This is evident from
logs.

    Error getting manifests for app ..., revision ...:
	rpc error: code = DeadlineExceeded desc = context deadline exceeded"
    Error generating diff for component ...:
	rpc error: code = DeadlineExceeded desc = context deadline exceeded"
    Handling of PR event failed: err=getting diff information:
	rpc error: code = DeadlineExceeded desc = context deadline exceeded\n" ...

Increasing the timeout is a short-term fix to mitigate the issue and
allow users to get the diff while additional investigation is done to
figure out how we can address this better long-term.

[1] https://github.com/commercetools/k8s-gitops/pull/998