Remove 'WithStructuredMergeDiff'.
This was introduced in v1.16.0. After deploying this version to
staging, Telefonistka fails to generate a diff with the following error:
time="2025-03-31T16:53:29Z" level=fatal msg="Failed to establish connection to argocd-server.argocd-infra.svc.cluster.local:443: listen unix /tmp/argocd-KIcBGaxaHKnAoaiE.sock: bind: no such file or directory"

This was not an issue in local teting.

Need to establish is the cause is this option, or another update.
The plan is to remove WithSTructuredMergeDiff option, do another release to staging,
and see if the issue is resolved.