Refactor getting PR ref and SHA
The only use of GetRef and GetSHA is in the comment event handler.

This refactors its use in an effort to simplify; the comment handler
instead fetches the PR object directly once, and aborts if the issue is
not linked to a PR.

It then sets the ref and SHA based on the returned object values.