Add support for retriggering via PR comment (#115)
* Add support for retriggering via PR comment

* fixup! Add support for retriggering via PR comment

* fixup! Add support for retriggering via PR comment

* Make handleChangedPREvent able to handle comment event

Instead of passing in the PullRequestEvent, the signature is changed to
accept the only used attributes of PR number and the list of labels.

This allows it to be called when receiving a IssueCommentEvent,
triggered from a "retrigger"-comment.

* Move panic handler

A triggering comment can now potentially cause a panic so the handler is
moved to cover such cases, avoiding a crashed server if there are bugs.

* Refactor getting PR ref and SHA

The only use of GetRef and GetSHA is in the comment event handler.

This refactors its use in an effort to simplify; the comment handler
instead fetches the PR object directly once, and aborts if the issue is
not linked to a PR.

It then sets the ref and SHA based on the returned object values.

* Trim off leading and trailing space

* Lint

* Log when bailing