Check in generated mocks (#126)
* Check in generated mocks

All Dependabot pull requests fail checks as the go mod tidy -diff step
fails. This is because tidy runs and sees dependencies from generated
files that has not been generated. That is:

    rm internal/pkg/mocks/argocd_*
    go mod tidy -diff

will fail showing a diff including the google.golang.org/grpc
dependency.

Currently there are two clear options.

1. Generate the files before running go mod tidy; or
2. Check in the generated files.

Normally the preference would be (1) to avoid including code not
actually written by developers maintaining a repository, but with that
said, it is customary in Go to include generated files so a project
builds successfully without an additional pre-compile step [1].

This was very elusive to track down but was ultimately confirmed with

    go mod why google.golang.org/grpc

pointing to the mocks package.

This change removes the (inconsistent) ignores of the generated files
and checks them in to avoid issues running go mod tidy -diff.