Build binary outside Dockerfile (#121)
* Trigger on workflow_dispatch

By adding workflow_dispatch as a trigger it is possible to execute the
workflow using the Github CLI. Without it, does not seem it is possible
to run on a non-main branch.

* Build outside Dockerfile

The build takes a ton of time and this attempts to improve the status
quo by building the binary outside of Docker.

By adding Go build steps and caching of dependencies and build artifacts
build time is reduced from about 10m to 1m.